moved the specialized iterator inside the _impl feature
